+++ .gitignore
@@ -12,3 +12,4 @@
 .idea/modules/aiku-common/aiku.aiku-common.main.iml
 .idea/modules/aiku-main/aiku.aiku-main.main.iml
 .idea/uiDesigner.xml
+.idea/modules/aiku-login/aiku.aiku-login.main.iml
